feat: New configurable `get_client_ip` (for rate-limiting & more)
Instead of using a default of XFF[0] or XFF[-1] or REMOTE_ADDR[0], all of
which are certainly wrong for someone, make IP determination customizable.

This replaces almost all uses of the django-ipware package. We weren't
using that package's XFF-chain-reading capabilities, and in any case it
isn't capable of handling trusted-proxy IP ranges in CIDR or other formats
that would be required for handling e.g. Cloudflare IPs (instead requiring
textual prefixes).

Related changes:

- Fix test that had wrong idea of XFF
- Give devstack a setting of REMOTE_ADDR[-1]

Also:

- Allow some safe-looking isort auto-fixes